+++ src/FREffects.hpp
@@ -1,15 +1,5 @@
 #include "TableNode.hpp"
 
-class FireSprite : public CCSprite {
-protected:
-    int m_currentFrame = 1;
-    GJFeatureState m_featureState;
-public:
-    static FireSprite* create(GJFeatureState);
-    bool init(GJFeatureState);
-    void fireAnimation(float);
-};
-
 class FREffects {
 public:
     static CCSprite* grdInfinity();